requirements.txt
spacy
datasets
lingua-language-detector
nltk